Allow anisotropic terms to be fit to dispersion term.
Anisotropic dispersion is currently assumed to be of the functional form

(cn/r^n)*f_n(x)*(1 + a*y10 + ...)

where f_n(x) is the usual Tang-Toennies damping factor and the last term is an
expansion in terms of spherical harmonics. Since the isotropic term is already
given by cn (which is read in as input), anisotropic dispersion has one less
free parameter compared to exchange.

In terms of implementation, the dispersion subroutines have changed somewhat
to accomodate anisotropic dispersion. If no anisotropic terms are given,
for efficiency dispersion is evaluated using numpy subroutines without calling
sympy. If anisotropic atomtypes are present, then sympy is called to evaluate
the dispersion function and derivatives, and parameter optimization is carried
out identically to exchange.