Create lambda functions for each atomtype interaction separately.
Previously, one lambda was generated for the entire force field energy as a
whole. The resulting lambda function is marginally faster than using several
lambda functions, but generating the function required a substantial amount of
overhead for large molecules (i.e. benzene dimer). Especially for molecules
with many atoms and few atomtypes (again, benzene is a good example), it's
much faster to generate and combine several lambda functions. This faster
method for generating the FF energy is now implemented. Sympy is still
used to get the interaction energies and derivatives for each pairwise
interaction energy.