Add support for multiple exponents per atomtype.
To deal with cases like Br2, where multiple Slater exponentials are required
to describe the local atomic density, and in order to eventually explore the
universality of the K parameter used to fit the overlap model to the exchange
energy, support was added to treat multiple exponents as well as to constrain
the A parameter fits based on the magnitude of the atomic densities
(D in p(r) = Dexp(-br)).

This support necessitated a restructuring of how fitting parameters are stored
internally in the program. Previously, free parameters were stored in lists;
now, both free and constrained paramters are stored in a list of dictionaries
(with each atomtype having its own dictionary).

Other minor changes included in this commit include a slight restructuring of
the output files for clarity.