Enable multipolar energy to be calculated for flexible monomer systems (i.e. where some/all of the monomer geometries in the .sapt file don't match the monomer geometry specified in the .mom files).
The old get_multipole_electrostatic_energy subroutine relied on calculating a
rotation matrix between the global (Cartesian) and local (multipolar)
coordinates, which in turn relied on being able to exactly rotate a monomer
(with coordinates specified by the .mom file) into the coordinates specified in
the .sapt file. This algorithm led to errors in cases where the internal
coordinates between .sapt and .mom files didn't match.

The new algorithm relies only on the local coordinate system for each atom, as
specified in the .axes file. Assuming this is properly specified, the new
algorithm works by rotating the *multipole moments themselves* into the local
coordinate frame for later use in multipolar energy calculations.

The current code works for calculating multipole energies, but future commits
will be needed to enable functionality for drude oscillator calculations, and
in general to clean up both the multipoles.py and rotations.py codes.